requirements.txt
Flask==0.10.1
requests==2.2.1
Pillow==3.1.0
